Revert "kermel: Refactoring and separating dynamic signature-related structures and functions and implement hot-signature updates"
This reverts commit a9e3c1cc8f410a20ba9f6fd3a1d2e34804355316.

Why would I revert this?
Because it will cause some nongki kernel panic.
