+++ kernel/core_hook.c
@@ -46,6 +46,7 @@
 #include "manager.h"
 #include "selinux/selinux.h"
 #include "throne_tracker.h"
+#include "throne_tracker.h"
 #include "kernel_compat.h"
 #include "kpm/kpm.h"
 #include "include/ksu_creds.h"
@@ -423,6 +424,7 @@ int ksu_handle_prctl(int option, unsigned long arg2, unsigned long arg3,
 				on_post_fs_data();
 				// Initializing Dynamic Signatures
         		ksu_dynamic_sign_init();
+        		ksu_load_dynamic_sign();
         		pr_info("Dynamic sign config loaded during post-fs-data\n");
 			}
 			break;
